move origin commands inside commands directory
